Fix for Chrome v21
Google Chrome v21+ requires a permission in order to download multiple
files at once, so the download mechanism has been refactored.
Also, code that is injected in the main page is now in a separate
namespace and shouldn't interfere with anything on the page in the
future.
Performed some more minor refactoring.