Fixed error with document.styleSheets[i].cssRules
I noticed that a lot of the sites that I'm trying to use the Image Downloader on have their stylesheets hosted on a CDN, which is a different domain. Due to cross-domain rules in Chrome, those stylesheets don't appear to be readable anymore, for security purposes.

Adding the try/catch block will mean that the extension shouldn't break, but it does mean that users will miss out on images that are defined in a stylesheet hosted on a CDN (or another domain entirely).